+++ src/ArduinoJson/Array/ArrayImpl.hpp
@@ -19,4 +19,10 @@ inline ObjectRef ArrayShortcuts<TArray>::createNestedObject() const {
   return impl()->addElement().template to<ObjectRef>();
 }
 
+template <typename TArray>
+inline ElementProxy<TArray> ArrayShortcuts<TArray>::operator[](
+    size_t index) const {
+  return ElementProxy<TArray>(*impl(), index);
+}
+
 }  // namespace ARDUINOJSON_NAMESPACE